This branch is "Fond Dream Finale" by @PalAlexx

+++ text.h
@@ -1,14 +1,15 @@
 /* -*- mode: c; tab-width: 4; c-basic-offset: 4; c-file-style: "linux" -*- */
 //
 // Copyright (c) 2009-2011, Wei Mingzhi <whistler_wmz@users.sf.net>.
-// Copyright (c) 2011-2023, SDLPAL development team.
+// Copyright (c) 2011-2019, SDLPAL development team.
 // All rights reserved.
 //
 // This file is part of SDLPAL.
 //
 // SDLPAL is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License, version 3
-// as published by the Free Software Foundation.
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
 //
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -108,7 +109,7 @@ PAL_GetMsgNum(
 
 LPWSTR
 PAL_UnescapeText(
-   LPCWSTR    lpszText
+	LPCWSTR    lpszText
 );
 
 VOID
@@ -123,13 +124,13 @@ PAL_DrawText(
 
 VOID
 PAL_DrawTextUnescape(
-   LPCWSTR    lpszText,
-   PAL_POS    pos,
-   BYTE       bColor,
-   BOOL       fShadow,
-   BOOL       fUpdate,
-   BOOL       fUse8x8Font,
-   BOOL       fUnescape
+	LPCWSTR    lpszText,
+	PAL_POS    pos,
+	BYTE       bColor,
+	BOOL       fShadow,
+	BOOL       fUpdate,
+	BOOL       fUse8x8Font,
+	BOOL       fUnescape
 );
 
 VOID
@@ -200,9 +201,9 @@ INT
 PAL_MultiByteToWideCharCP(
 	CODEPAGE      cp,
 	LPCSTR        mbs,
-	size_t        mbslength,
+	int           mbslength,
 	LPWSTR        wcs,
-	size_t        wcslength
+	int           wcslength
 	);
 
 WCHAR
@@ -223,7 +224,7 @@ PAL_SetCodePage(
 CODEPAGE
 PAL_DetectCodePageForString(
 	const char *   text,
-	size_t         text_len,
+	int            text_len,
 	CODEPAGE       default_cp,
 	int *          probability
 );
